% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis-function.R
\name{diaglm}
\alias{diaglm}
\title{test whether linear model assumptions have been violated}
\usage{
diaglm(x, file = "result.txt")
}
\arguments{
\item{x}{an object, which is a result from a linear model (lm)}

\item{file}{is the name of text file you want to create (e.g.,'mydiagnosis.txt'). It is iniated as 'result.txt'
Note that the file will be saved at the working directory
data(data)
mod<-lm(data$tolerance_scale~data$sg_participation+data$virtual_si+as.numeric(data$EDUC)+as.numeric(data$yrbrn)+data$IVRACE2+data$IVRACE3+data$gndr+as.numeric(data$lrscale))
diaglm(mod,data$tolerance_scale)}
}
\value{
The statistical tests results, outlierTest, durbinWatsonTest, VIF, and gvlma
}
\description{
test for unusual data, nonlinearity, non-normality, non-constant error variance, auto-correlation
}
