% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis-function.R
\name{diagns}
\alias{diagns}
\title{test whether linear model assumptions have been violated}
\usage{
diagns(x, y)
}
\arguments{
\item{x}{an object, which is a result from a linear model}

\item{y}{is the name of text file you want to create (e.g.,'mydiagnosis.txt')
Note that the file will be saved at the working directory}
}
\value{
The statistical tests results, outlierTest, durbinWatsonTest, VIF, and gvlma
}
\section{Functions}{
\itemize{
\item \code{diagns}: for unusual data, nonlinearity, non-normality, non-constant error variance, auto-correlation
}}

