% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformX-nocontrol.R
\name{transformX.nocontrol}
\alias{transformX.nocontrol}
\title{find the best transformatio of explanatory variables in multiple regression}
\usage{
transformX.nocontrol(mod, d, file = "transformX.txt")
}
\arguments{
\item{mod}{a formula like y~X1+X2 for the linear model}

\item{file}{the default is'transformX.txt'
Note that the file will be saved at the working directory}

\item{data}{the dataset you will run on}
}
\value{
The statistical tests results based on the Maximum likelihood
}
\description{
tranaform the X variables. Different from transformX because here all variables are considered for transformation
}
\examples{
data(data)
transformX.nocontrol(tolerance_scale~as.numeric(EDUC)+as.numeric(lrscale),data)
}
