% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformX.R
\name{transformX}
\alias{transformX}
\title{find the best transformatio of explanatory variables in multiple regression}
\usage{
transformX(mod, control, d, file = "transformX.txt")
}
\arguments{
\item{mod}{a formula like y~X1+X2 for the linear model}

\item{control}{the control variable(s) are those in the model that you don't want to transform, use the form of ~x3+x4}

\item{file}{the default is'transformX.txt'
Note that the file will be saved at the working directory}

\item{data}{the dataset you will run on}
}
\value{
The statistical tests results based on the Maximum likelihood
}
\description{
tranaform the X variables in multiple regression
}
\examples{
data(data)
transformX(tolerance_scale~as.numeric(EDUC)+as.numeric(lrscale),~factor(gndr),data)
}
