% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{virtual}
\alias{virtual}
\title{virtual. dataset for replication study of Kittilson and Dalton (2011)
Original article used virtual. Kittilson, Miki Caul and Russell J. Dalton. (2011). “Virtual Civil Society: The New Frontier of Social Capital?” Political Behavior: Volume 33, Issue 4, pp 625-644}
\format{A data frame with 1001 rows and 804 variables}
\source{
\url{https://www.icpsr.umich.edu/icpsrweb/ICPSR/studies/4607}
}
\usage{
data(virtual)
}
\description{
virtual. The newly added variables are as follows:
\itemize{
  \item tolerance_scale. the response variable for model 1 of Table 4, Kittilson and Dalton (2011)
  \item duty_based_index. the response variable for model 2 of Table 4, Kittilson and Dalton (2011)
  \item engagement_index. the response variable for model 3 of Table 4, Kittilson and Dalton (2011)
  \item Factor1. the response variable for model 5 of Table 4, Kittilson and Dalton (2011)
  \item Factor2. the response variable for model 6 of Table 4, Kittilson and Dalton (2011)
  \item Factor3. the response variable for model 7 of Table 4, Kittilson and Dalton (2011)
  \item Factor4. the response variable for model 1 of Table 4, Kittilson and Dalton (2011)
}
}
\keyword{datasets}
