% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_response_variable.R
\name{transformY}
\alias{transformY}
\title{find the best transformatio of response variable, and then run the linear model based on the best transformation}
\usage{
transformY(x, y, file = "tranformResult.txt")
}
\arguments{
\item{x}{an lm object, which is a result from a linear model (lm)}

\item{y}{the response variable you would like to transform}

\item{file}{is the name of text file you want to create (e.g.,'mydiagnosis.txt'). It is iniated as 'result.txt'
Note that the file will be saved at the working directory}
}
\value{
The statistical tests results, outlierTest, durbinWatsonTest, VIF, and gvlma

y1round the transformed y vector as in the global environment
}
\description{
tranaform the y and rerun the linear model
}
